# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "ApiKeyListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ApiKeyResponse"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "ApiKeyListResponse",
        "type": "object"
      },
...
      "CreateApiKeyRequest": {
        "additionalProperties": false,
        "properties": {
          "expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/Role",
            "default": "developer",
            "description": "Cannot exceed your own role."
          },
          "scopes": {
            "description": "Optional narrowing. Empty means the role's full permissions.",
            "items": {
              "$ref": "#/components/schemas/Permission"
            },
            "title": "Scopes",
            "type": "array"
          }
        },
        "required": [
          "name"
        ],
        "title": "CreateApiKeyRequest",
        "type": "object"
      },
...
      "CreateOrganisationRequest": {
        "additionalProperties": false,
        "properties": {
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "CreateOrganisationRequest",
        "type": "object"
      },
...
      "CreatedApiKeyResponse": {
        "description": "Returned once, at creation.\n\n``token`` is the only time the full credential exists anywhere; the platform\nstores only its SHA-256 digest and cannot show it again.",
        "properties": {
          "key": {
            "$ref": "#/components/schemas/ApiKeyResponse"
          },
          "token": {
            "description": "Copy this now. It cannot be retrieved later.",
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "key",
          "token"
        ],
        "title": "CreatedApiKeyResponse",
        "type": "object"
      },
...
      "OrganisationListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/OrganisationResponse"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "OrganisationListResponse",
        "type": "object"
      },
...
      "OrganisationResponse": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "created_at"
        ],
        "title": "OrganisationResponse",
        "type": "object"
      },
...
      "ProblemDetail": {
        "description": "The body returned for every non-2xx API response.",
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode",
            "description": "Stable machine-readable error code."
          },
          "correlation_id": {
            "description": "Quote this when reporting a failure.",
            "title": "Correlation Id",
            "type": "string"
          },
          "details": {
            "description": "Structured, non-sensitive context such as failing field paths.",
            "title": "Details",
            "type": "object"
          },
          "message": {
            "description": "Human-readable summary. Safe to display.",
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "correlation_id"
        ],
        "title": "ProblemDetail",
        "type": "object"
      },
...
      "Role": {
        "description": "A member's role within one organisation.",
        "enum": [
          "owner",
          "admin",
          "developer",
          "reviewer",
          "viewer"
        ],
        "title": "Role",
        "type": "string"
      },
...
    }
  },
  "info": {
    "description": "AgentRail platform API.\n\nPhase 0 exposes the deterministic job slice used to prove the end-to-end request\npath: the API records a job in PostgreSQL, publishes its identifier to Redis, and\na worker executes it against the CloudOps sandbox.",
    "title": "AgentRail API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/organisations": {
      "get": {
        "operationId": "list_organisations",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganisationListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Not signed in."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Not yours, or not permitted."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Already exists."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Validation failed."
          }
        },
        "summary": "List organisations the caller belongs to",
        "tags": [
          "organisations"
        ]
      },
      "post": {
        "operationId": "create_organisation",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrganisationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganisationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Not signed in."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Not yours, or not permitted."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Already exists."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Validation failed."
          }
        },
        "summary": "Create an organisation",
        "tags": [
          "organisations"
        ]
      }
    },
    "/api/v1/organisations/{organisation_id}": {
      "get": {
        "operationId": "get_organisation",
        "parameters": [
          {
            "in": "path",
            "name": "organisation_id",
            "required": true,
            "schema": {
              "maxLength": 26,
              "minLength": 26,
              "title": "Organisation Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganisationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Not signed in."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Not yours, or not permitted."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Already exists."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Validation failed."
          }
        },
        "summary": "Fetch one organisation",
        "tags": [
          "organisations"
        ]
      }
    },
    "/api/v1/organisations/{organisation_id}/api-keys": {
      "get": {
        "operationId": "list_api_keys",
        "parameters": [
          {
            "in": "path",
            "name": "organisation_id",
            "required": true,
            "schema": {
              "maxLength": 26,
              "minLength": 26,
              "title": "Organisation Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Not signed in."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Not yours, or not permitted."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Already exists."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Validation failed."
          }
        },
        "summary": "List API keys (never includes secrets)",
        "tags": [
          "organisations"
        ]
      },
      "post": {
        "description": "The response contains the only copy of the token that will ever exist.",
        "operationId": "create_api_key",
        "parameters": [
          {
            "in": "path",
            "name": "organisation_id",
            "required": true,
            "schema": {
              "maxLength": 26,
              "minLength": 26,
              "title": "Organisation Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateApiKeyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedApiKeyResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Not signed in."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Not yours, or not permitted."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Already exists."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Validation failed."
          }
        },
        "summary": "Create a scoped API key",
        "tags": [
          "organisations"
        ]
      }
    },
    "/api/v1/organisations/{organisation_id}/api-keys/{key_id}": {
      "delete": {
        "operationId": "revoke_api_key",
        "parameters": [
          {
            "in": "path",
            "name": "organisation_id",
            "required": true,
            "schema": {
              "maxLength": 26,
              "minLength": 26,
              "title": "Organisation Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "key_id",
            "required": true,
            "schema": {
              "maxLength": 26,
              "minLength": 26,
              "title": "Key Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Not signed in."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Not yours, or not permitted."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Already exists."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Validation failed."
          }
        },
        "summary": "Revoke an API key",
        "tags": [
          "organisations"
        ]
      }
    },
    "/api/v1/organisations/{organisation_id}/audit-events": {
      "get": {
        "operationId": "list_audit_events",
        "parameters": [
          {
            "in": "path",
            "name": "organisation_id",
            "required": true,
            "schema": {
              "maxLength": 26,
              "minLength": 26,
              "title": "Organisation Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditEventListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Not signed in."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Not yours, or not permitted."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Already exists."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Validation failed."
          }
        },
        "summary": "Recent audit events",
        "tags": [
          "organisations"
        ]
      }
    },
    "/api/v1/organisations/{organisation_id}/members": {
      "get": {
        "operationId": "list_members",
        "parameters": [
          {
            "in": "path",
            "name": "organisation_id",
            "required": true,
            "schema": {
              "maxLength": 26,
              "minLength": 26,
              "title": "Organisation Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Not signed in."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Not yours, or not permitted."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Already exists."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Validation failed."
          }
        },
        "summary": "List members",
        "tags": [
          "organisations"
        ]
      },
      "post": {
        "operationId": "add_member",
        "parameters": [
          {
            "in": "path",
            "name": "organisation_id",
            "required": true,
            "schema": {
              "maxLength": 26,
              "minLength": 26,
              "title": "Organisation Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddMemberRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Not signed in."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Not yours, or not permitted."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Already exists."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Validation failed."
          }
        },
        "summary": "Grant a role to an existing user",
        "tags": [
          "organisations"
        ]
      }
    },
...
  }
}